# src/test/resources/contract/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
      "ExamDto" : {
        "description" : "Data Transfer Object representing the official structure of the exam (e.g., from an edital).",
        "properties" : {
          "examDate" : {
            "description" : "The scheduled date for the exam. Must be in the future.",
            "example" : "2024-12-01",
            "format" : "date",
            "type" : "string"
          },
          "generalKnowledgeSubjects" : {
            "description" : "List of subjects evaluated in the general knowledge section.",
            "items" : {
              "$ref" : "#/components/schemas/SubjectDto"
            },
            "maxItems" : 50,
            "minItems" : 0,
            "type" : "array"
          },
          "generalKnowledgeTotalScore" : {
            "description" : "The total score available in the general knowledge section.",
            "example" : 50.0,
            "format" : "double",
            "maximum" : 1000.0,
            "minimum" : 0,
            "type" : "number"
          },
          "name" : {
            "description" : "The official name of the exam.",
            "example" : "Federal Police Officer 2024",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "specificKnowledgeAxes" : {
            "description" : "List of thematic axes, which group specific knowledge subjects and define their weights.",
            "items" : {
              "$ref" : "#/components/schemas/ThematicAxisDto"
            },
            "maxItems" : 50,
            "minItems" : 0,
            "type" : "array"
          }
        },
        "required" : [ "examDate", "generalKnowledgeSubjects", "name", "specificKnowledgeAxes" ],
        "type" : "object"
      },
...
      "GaConfigDto" : {
        "description" : "Configuration settings for the Genetic Algorithm.",
        "properties" : {
          "numGenerations" : {
            "description" : "The number of evolutionary cycles the algorithm will run. Higher yields better results but takes longer.",
            "example" : 100,
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 10,
            "type" : "integer"
          },
          "populationSize" : {
            "description" : "The number of potential solutions evolving per generation.",
            "example" : 50,
            "format" : "int32",
            "maximum" : 500,
            "minimum" : 10,
            "type" : "integer"
          },
          "randomSeed" : {
            "description" : "Seed of the pseudo-random generator. Supply it to make a run reproducible: the same payload with the same seed yields the same plan, day for day. Omit it for an unseeded run, which is the previous behaviour.",
            "example" : 20260903,
            "format" : "int64",
            "type" : "integer"
          },
          "totalStudyDays" : {
            "description" : "The target number of study days to allocate across all subjects.",
            "example" : 100,
            "format" : "int32",
            "maximum" : 365,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "JobAcceptedDto" : {
        "description" : "Confirmation that the optimization request was accepted for processing.",
        "properties" : {
          "id" : {
            "description" : "Identifier of the accepted job.",
            "example" : "3f2b1c44-9a7e-4c1f-8f0d-2a5b6c7d8e9f",
            "type" : "string"
          },
          "status" : {
            "description" : "Current status of the job.",
            "example" : "PENDING",
            "type" : "string"
          },
          "statusUrl" : {
            "description" : "Where to poll for the result.",
            "example" : "/api/v1/optimizer/jobs/3f2b1c44-9a7e-4c1f-8f0d-2a5b6c7d8e9f",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "JobStatusDto" : {
        "description" : "Status of an optimization job, including the result once it is ready.",
        "properties" : {
          "error" : {
            "description" : "Why it failed. Present only when FAILED.",
            "type" : "string"
          },
          "finishedAt" : {
            "description" : "When it finished. Absent while not finished.",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "Job identifier.",
            "type" : "string"
          },
          "result" : {
            "$ref" : "#/components/schemas/PlannerResponseDto",
            "description" : "The generated plan. Present only when SUCCEEDED."
          },
          "startedAt" : {
            "description" : "When it left the queue and started running. Absent while PENDING.",
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "description" : "PENDING, RUNNING, SUCCEEDED or FAILED.",
            "example" : "SUCCEEDED",
            "type" : "string"
          },
          "submittedAt" : {
            "description" : "When the request was accepted.",
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "OptimizationRequest" : {
        "description" : "The main request payload for generating an optimized study plan.",
        "properties" : {
          "exam" : {
            "$ref" : "#/components/schemas/ExamDto",
            "description" : "The structure and rules of the exam."
          },
          "gaConfig" : {
            "$ref" : "#/components/schemas/GaConfigDto",
            "description" : "Configuration parameters for the underlying Genetic Algorithm engine."
          },
          "studentProfile" : {
            "$ref" : "#/components/schemas/StudentProfileDto",
            "description" : "The personal profile of the student, including availability and knowledge gaps."
          }
        },
        "required" : [ "exam", "gaConfig", "studentProfile" ],
        "type" : "object"
      },
      "OptimizationResultDto" : {
        "description" : "The strategic result generated by the Genetic Algorithm.",
        "properties" : {
          "coreVersion" : {
            "description" : "Version of the build that produced this plan. Together with the request payload and its randomSeed, it is what makes a run replayable.",
            "example" : "2.0.1",
            "type" : "string"
          },
          "executionTimeMillis" : {
            "description" : "The total execution time of the optimization algorithm in milliseconds.",
            "example" : 45,
            "format" : "int64",
            "type" : "integer"
          },
          "fitness" : {
            "description" : "The final calculated fitness score of the best individual.",
            "example" : 142.55,
            "format" : "double",
            "type" : "number"
          },
          "fitnessBreakdown" : {
            "$ref" : "#/components/schemas/FitnessBreakdown",
            "description" : "The same fitness, decomposed into the terms that produced it: each term's normalised value, its declared weight and what it contributed. The aggregate is exactly the fitness field above."
          },
          "generationsRun" : {
            "description" : "The number of evolutionary generations processed before finding this solution.",
            "example" : 100,
            "format" : "int32",
            "type" : "integer"
          },
          "plan" : {
            "$ref" : "#/components/schemas/StudyPlanDto",
            "description" : "The allocated study days for each subject."
          }
        },
        "type" : "object"
      },
...
      "PlannerResponseDto" : {
        "description" : "The comprehensive response containing both the strategic and tactical study plans.",
        "properties" : {
          "message" : {
            "description" : "A status message summarizing the generation outcome.",
            "example" : "Full study plan generated successfully.",
            "type" : "string"
          },
          "optimizationResult" : {
            "$ref" : "#/components/schemas/OptimizationResultDto",
            "description" : "The strategic output from the Genetic Algorithm (e.g., total days per subject)."
          },
          "scheduleResult" : {
            "$ref" : "#/components/schemas/ScheduleResultDto",
            "description" : "The tactical, day-by-day study schedule based on the optimization result."
          }
        },
        "type" : "object"
      },
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "format" : "uri",
            "type" : [ "string", "null" ]
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : [ "object", "null" ]
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "StudentProfileDto" : {
        "description" : "Data Transfer Object representing the student's personal constraints and knowledge state.",
        "properties" : {
          "knowledgeGaps" : {
            "additionalProperties" : {
              "format" : "double",
              "type" : "number"
            },
            "description" : "A mapping of subject names to the student's self-assessed knowledge gap on a 1.0 (very strong) to 5.0 (very weak) scale.",
            "example" : {
              "Constitutional Law" : 4.5,
              "Portuguese" : 2.0
            },
            "minProperties" : 1,
            "type" : "object"
          },
          "name" : {
            "description" : "The name of the student.",
            "example" : "Jane Doe",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "state" : {
            "$ref" : "#/components/schemas/StudentStateDto",
            "description" : "The psychological, physical, and emotional state of the student."
          },
          "weeklyAvailability" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "description" : "A mapping of the days of the week to the number of hours the student is available to study on that day.",
            "example" : {
              "MONDAY" : 4,
              "SATURDAY" : 8
            },
            "minProperties" : 1,
            "type" : "object"
          }
        },
        "required" : [ "knowledgeGaps", "name", "weeklyAvailability" ],
        "type" : "object"
      },
...
      "StudyPlanDto" : {
        "description" : "The core output of the Genetic Algorithm, representing the optimal distribution of study days across subjects.",
        "properties" : {
          "daysPerSubject" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "description" : "A mapping where keys are subject names and values are the optimal number of days to study them.",
            "example" : {
              "Constitutional Law" : 10,
              "Portuguese" : 5
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "contact" : {
      "email" : "engineering@example.com",
      "name" : "Engineering Team"
    },
    "description" : "A stateless REST API that uses Genetic Algorithms to orchestrate an optimized and mathematically superior study plan. Built with Spring Boot.",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "Dynamic Study Planner API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/optimizer/generate" : {
      "post" : {
        "description" : "Runs a Genetic Algorithm to find the optimal allocation of study days, then generates a day-by-day tactical schedule.",
        "operationId" : "generateFullStudyPlan",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlannerResponseDto"
                }
              }
            },
            "description" : "Successfully generated the optimized study plan."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request. The payload could not be understood: failed validation, malformed JSON, or a field with an incompatible type."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found. No endpoint is mapped to the requested path."
          },
          "405" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Method Not Allowed. This endpoint accepts POST only; the response carries an Allow header."
          },
          "408" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Request Timeout. The algorithm took too long to compute."
          },
          "415" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unsupported Media Type. The request body must be application/json."
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content. The request was understood but cannot be fulfilled — typically the subjects require more days than remain before the exam date."
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Too Many Requests. The per-client rate limit was exceeded."
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Internal Server Error. An unexpected exception occurred."
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Service Unavailable. The service is at capacity and refused the request, or a required dependency is unavailable. Retry after the number of seconds in the Retry-After header."
          }
        },
        "summary" : "Generate an optimized study plan",
        "tags" : [ "Study Plan Optimizer" ]
      }
    },
    "/api/v1/optimizer/jobs" : {
      "post" : {
        "description" : "Accepts the request, returns an identifier immediately and runs the genetic algorithm in the background. Poll the returned URL for the result.",
        "operationId" : "submeter",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobAcceptedDto"
                }
              }
            },
            "description" : "Accepted. The Location header and statusUrl point to the job."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request. The payload failed validation or could not be parsed."
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Too Many Requests. The per-client rate limit was exceeded."
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Service Unavailable. The queue is full; retry after the number of seconds in the Retry-After header."
          }
        },
        "summary" : "Submit an optimization request",
        "tags" : [ "Study Plan Optimizer (asynchronous)" ]
      }
    },
    "/api/v1/optimizer/jobs/{id}" : {
      "get" : {
        "operationId" : "consultar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobStatusDto"
                }
              }
            },
            "description" : "The job exists. The result field is present once SUCCEEDED."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No job with this identifier — never existed, or already expired."
          }
        },
        "summary" : "Read the status and, once ready, the result of a job",
        "tags" : [ "Study Plan Optimizer (asynchronous)" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Local Development Environment",
    "url" : "http://localhost:8080"
  } ],
...
}